update 7 files

+++ src/main/java/hexlet/code/controller/UserController.java
@@ -5,6 +5,7 @@ import hexlet.code.exception.UserNotFoundException;
 import hexlet.code.model.User;
 import hexlet.code.model.UserRole;
 import hexlet.code.repository.UserRepository;
+import hexlet.code.service.UserDetailsServiceImpl;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -21,6 +22,12 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 
 
 import java.util.List;
+import java.util.Objects;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @RequestMapping("/api/users")
@@ -32,17 +39,33 @@ public class UserController {
     @Autowired
     private PasswordEncoder encoder;
 
+    @Autowired
+    private UserDetailsServiceImpl userDetailsServiceImpl;
+
+    @Operation(summary = "Get all users")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Get list of all users")
+    })
     @GetMapping(path = "")
     public List<User> getUsers() {
         return userRepository.findAll();
     }
 
+    @Operation(summary = "Get user by id")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Get a specific user by id")
+    })
     @GetMapping(path = "/{id}")
     public User getUser(@PathVariable long id) {
         return userRepository.findById(id)
                .orElseThrow(() -> new UserNotFoundException("User" + id + "not found"));
     }
 
+    @Operation(summary = "Create new user")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "User successfully created"),
+            @ApiResponse(responseCode = "422", description = "Data not valid")
+    })
     @PostMapping(path = "")
     @ResponseStatus(HttpStatus.CREATED)
     public User createUser(@Valid @RequestBody UserDto userDto) {
@@ -56,12 +79,23 @@ public class UserController {
         return userRepository.save(user);
     }
 
+    @Operation(summary = "Update user by id")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "User successfully updated"),
+            @ApiResponse(responseCode = "422", description = "Data not valid")
+    })
     @PutMapping(path = "/{id}")
     public User updateUser(@Valid @RequestBody UserDto userDto, @PathVariable Long id) {
         if (!userRepository.existsById(id)) {
             throw new UserNotFoundException("User" + id + "not found");
         }
         User user = userRepository.findById(id).get();
+
+        User currentUser = userDetailsServiceImpl.getCurrentUserName();
+
+        if (!Objects.equals(currentUser, user)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
 
         user.setFirstName(userDto.getFirstName());
         user.setLastName(userDto.getLastName());
@@ -71,11 +105,20 @@ public class UserController {
         return userRepository.save(user);
     }
 
+    @Operation(summary = "Delete user by id")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Delete user by id")
+    })
     @DeleteMapping(path = "/{id}")
     public void deleteUser(@PathVariable long id) {
         if (!userRepository.existsById(id)) {
-            // Если не существует, возвращаем код ответа 404
             throw new UserNotFoundException("User" + id + "not found");
+        }
+        User user = userRepository.findById(id).get();
+        User currentUser = userDetailsServiceImpl.getCurrentUserName();
+
+        if (!Objects.equals(currentUser, user)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         }
         userRepository.deleteById(id);
     }
